pyproject.toml
[project]
name = "mini-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "grandalf>=0.8",
    "ipython>=8.37.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.0",
    "langchain-milvus>=0.2.2",
    "langchain-openai>=1.0.1",
    "langchain-tavily>=0.2.12",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "notebook>=7.4.7",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
